Revise the complete library
Make the whole library simpler, by using template specialisation less and functions more.
This is a breaking change, because the interface of many functions has changed.
This allows directions to be the last parameter.
This is more natural in C++, since only the last parameters can have defaults.

This does not compile on GCC 4.6 any more.